Commit of TinyScheme version 1.29
The previous version contained a lot of corrections, but there were a lot
more that still wait on a sheet of paper lost in a carton someplace after
my house move... Manuel Heras-Gilsanz noticed this and resent his own
contribution, which relies on another bugfix that v.1.28 was missing: a
problem with string output, that this version fixes. I hope other people
will take the time to resend their contributions, if they didn't make it
to v.1.28.


git-svn-id: https://tinyscheme.svn.sourceforge.net/svnroot/tinyscheme/trunk@12 4a0124f5-7288-4c88-8df5-efb20b5748ad