Added narg() FFI macro to get a number argument.
When the type of numeric Scheme argument does not matter (real/integer),
use the narg() macro to get the value, and it will be converted automatically
to the propper type, instead of complaining about type mismatch.

Example, suppose your function is called with 2 arguments, 1.37 and 5, then:

int n;
double d;

iarg(n); // FAILURE! Type mismatch! Not an integer, but a real!
rarg(d); // FAILURE! Type mismatch! Not a real, but an integer!

Instead, use the following:

narg(n); // SUCCESS! Automatically converted to 1.
narg(d); // SUCCESS! Automatically converted to 5.0.