Write and engage debian/Makefile that uses implicit rules for easier hardening.
debian/Makefile is invoked in place of the upstream makefile within
debian/rules.  A new stanza in debian/rules creates an executable
named tinyscheme which needs to be cleaned.  Installation is migrated
from the patched makefile to debian/install.  Quilt patch to upstream
makefile as it is unused, so has been deleted.